+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixes issue with print dht2 when multipliers are a data.frame (Issue #179)
 * Fixes bug when including a uniform with no adjustment terms in the summarize_ds_models function (Issue #180)
+* dht2 can now deal with strata where there are no observations (Issue #194)
 
 # Distance 2.0.0